+++ cms/static/css/nhsuk.css
@@ -8764,8 +8764,7 @@ nav a:hover, nav a:visited {
 }
 
 /*
-/*
-JS:        packages/mega-menu-expander.js
+JS:        packages/mega-menu.js
 CSS:       packages/custom-styles/_megamenu.scss
 Templates: cms/templates/partials/megamenu
 */
