+++ cms/static/js/nhsuk.js
@@ -229,25 +229,14 @@ eval("/**\n * IE polyfill for NodeList.forEach()\n */\nif (!NodeList.prototype.f
 
 /***/ }),
 
-/***/ "./packages/mega-menu-expander.js":
-/*!****************************************!*\
-  !*** ./packages/mega-menu-expander.js ***!
-  \****************************************/
+/***/ "./packages/mega-menu.js":
+/*!*******************************!*\
+  !*** ./packages/mega-menu.js ***!
+  \*******************************/
 /*! no static exports found */
 /***/ (function(module, exports) {
 
-eval("function _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== \"undefined\" && o[Symbol.iterator] || o[\"@@iterator\"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === \"number\") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError(\"Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it[\"return\"] != null) it[\"return\"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\n/*\nJS:        packages/mega-menu-expander.js\nCSS:       packages/custom-styles/_megamenu.scss\nTemplates: cms/templates/partials/megamenu\n*/\ndocument.addEventListener(\"DOMContentLoaded\", function () {\n  var toggleMenu = document.getElementsByClassName(\"toggle-menu-expand\");\n\n  var _iterator = _createForOfIteratorHelper(toggleMenu),\n      _step;\n\n  try {\n    for (_iterator.s(); !(_step = _iterator.n()).done;) {\n      var link = _step.value;\n      link.addEventListener(\"click\", function (event) {\n        // Add listener here\n        event.preventDefault();\n      });\n    }\n  } catch (err) {\n    _iterator.e(err);\n  } finally {\n    _iterator.f();\n  }\n});\n\n//# sourceURL=webpack:///./packages/mega-menu-expander.js?");
-
-/***/ }),
-
-/***/ "./packages/mega-menu-toggle.js":
-/*!**************************************!*\
-  !*** ./packages/mega-menu-toggle.js ***!
-  \**************************************/
-/*! no static exports found */
-/***/ (function(module, exports) {
-
-eval("var show = function show(elem) {\n  var getHeight = function getHeight() {\n    elem.style.display = 'block';\n    var height = elem.scrollHeight + 'px';\n    elem.style.display = '';\n    return height;\n  };\n\n  var height = getHeight();\n  elem.classList.add('is-visible');\n  elem.style.height = height;\n  window.setTimeout(function () {\n    elem.style.height = '';\n  }, 350);\n};\n\nvar hide = function hide(elem) {\n  elem.style.height = elem.scrollHeight + 'px';\n  window.setTimeout(function () {\n    elem.style.height = '0';\n  }, 1);\n  window.setTimeout(function () {\n    elem.classList.remove('is-visible');\n  }, 350);\n};\n\nvar toggle = function toggle(elem, timing) {\n  if (elem.classList.contains('is-visible')) {\n    hide(elem);\n    return;\n  }\n\n  show(elem);\n};\n\ndocument.getElementById(\"toggle-menu\").addEventListener(\"click\", function (e) {\n  e.preventDefault();\n  toggle(document.getElementById(\"mega-menu\"));\n}, false);\ndocument.addEventListener(\"DOMContentLoaded\", function () {\n  if (location.hash.replace('#', '') === \"open\") {\n    toggle(document.getElementById(\"mega-menu\"));\n  }\n});\n\n//# sourceURL=webpack:///./packages/mega-menu-toggle.js?");
+eval("function _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== \"undefined\" && o[Symbol.iterator] || o[\"@@iterator\"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === \"number\") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e) { throw _e; }, f: F }; } throw new TypeError(\"Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e2) { didErr = true; err = _e2; }, f: function f() { try { if (!normalCompletion && it[\"return\"] != null) it[\"return\"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\n/*\nJS:        packages/mega-menu.js\nCSS:       packages/custom-styles/_megamenu.scss\nTemplates: cms/templates/partials/megamenu\n*/\n\n/* ******************** Burger Menu ******************** */\nvar show = function show(elem) {\n  var getHeight = function getHeight() {\n    elem.style.display = 'block';\n    var height = elem.scrollHeight + 'px';\n    elem.style.display = '';\n    return height;\n  };\n\n  var height = getHeight();\n  elem.classList.add('is-visible');\n  elem.style.height = height;\n  window.setTimeout(function () {\n    elem.style.height = '';\n  }, 350);\n};\n\nvar hide = function hide(elem) {\n  elem.style.height = elem.scrollHeight + 'px';\n  window.setTimeout(function () {\n    elem.style.height = '0';\n  }, 1);\n  window.setTimeout(function () {\n    elem.classList.remove('is-visible');\n  }, 350);\n};\n\nvar toggle = function toggle(elem, timing) {\n  if (elem.classList.contains('is-visible')) {\n    hide(elem);\n    return;\n  }\n\n  show(elem);\n};\n\ndocument.getElementById(\"toggle-menu\").addEventListener(\"click\", function (e) {\n  e.preventDefault();\n  toggle(document.getElementById(\"mega-menu\"));\n}, false);\ndocument.addEventListener(\"DOMContentLoaded\", function () {\n  if (location.hash.replace('#', '') === \"open\") {\n    toggle(document.getElementById(\"mega-menu\"));\n  }\n});\n/* ******************** Mega Menu ******************** */\n\ndocument.addEventListener(\"DOMContentLoaded\", function () {\n  var toggleMenu = document.getElementsByClassName(\"toggle-menu-expand\");\n\n  var _iterator = _createForOfIteratorHelper(toggleMenu),\n      _step;\n\n  try {\n    for (_iterator.s(); !(_step = _iterator.n()).done;) {\n      var link = _step.value;\n      link.addEventListener(\"click\", function (event) {// Add listener here\n        // if (Number(event.target.getAttribute(\"data-depth\")) <= 2) {\n        //     event.preventDefault();\n        // }\n      });\n    }\n  } catch (err) {\n    _iterator.e(err);\n  } finally {\n    _iterator.f();\n  }\n});\n\n//# sourceURL=webpack:///./packages/mega-menu.js?");
 
 /***/ }),
 
@@ -259,7 +248,7 @@ eval("var show = function show(elem) {\n  var getHeight = function getHeight() {
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_nhsuk_frontend_packages_nhsuk__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../node_modules/nhsuk-frontend/packages/nhsuk */ \"./node_modules/nhsuk-frontend/packages/nhsuk.js\");\n/* harmony import */ var _search_toggle_fix__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./search-toggle-fix */ \"./packages/search-toggle-fix.js\");\n/* harmony import */ var _search_toggle_fix__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_search_toggle_fix__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _mega_menu_toggle__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./mega-menu-toggle */ \"./packages/mega-menu-toggle.js\");\n/* harmony import */ var _mega_menu_toggle__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(_mega_menu_toggle__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var _mega_menu_expander__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./mega-menu-expander */ \"./packages/mega-menu-expander.js\");\n/* harmony import */ var _mega_menu_expander__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(_mega_menu_expander__WEBPACK_IMPORTED_MODULE_3__);\n\n //for issue https://github.com/rkhleics/nhs-ei.website/issues/29\n\n\n\n\n//# sourceURL=webpack:///./packages/nhsuk.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_nhsuk_frontend_packages_nhsuk__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../node_modules/nhsuk-frontend/packages/nhsuk */ \"./node_modules/nhsuk-frontend/packages/nhsuk.js\");\n/* harmony import */ var _search_toggle_fix__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./search-toggle-fix */ \"./packages/search-toggle-fix.js\");\n/* harmony import */ var _search_toggle_fix__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_search_toggle_fix__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _mega_menu__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./mega-menu */ \"./packages/mega-menu.js\");\n/* harmony import */ var _mega_menu__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(_mega_menu__WEBPACK_IMPORTED_MODULE_2__);\n\n //for issue https://github.com/rkhleics/nhs-ei.website/issues/29\n\n\n\n//# sourceURL=webpack:///./packages/nhsuk.js?");
 
 /***/ }),
