+++ cms/templates/partials/megamenu/README
@@ -1,3 +1,3 @@
-JS:        packages/mega-menu-expander.js
+JS:        packages/mega-menu.js
 CSS:       packages/custom-styles/_megamenu.scss
 Templates: cms/templates/partials/megamenu
