+++ packages/custom-styles/_megamenu.scss
@@ -1,6 +1,5 @@
 /*
-/*
-JS:        packages/mega-menu-expander.js
+JS:        packages/mega-menu.js
 CSS:       packages/custom-styles/_megamenu.scss
 Templates: cms/templates/partials/megamenu
 */
