+++ packages/nhsuk.js
@@ -2,5 +2,4 @@ import '../node_modules/nhsuk-frontend/packages/nhsuk';
 
 import './search-toggle-fix'; //for issue https://github.com/rkhleics/nhs-ei.website/issues/29
 
-import './mega-menu-toggle';
-import './mega-menu-expander';
+import './mega-menu';
